Comapny Entity mapped to Jobs

+++ EmployEase/src/main/java/com/hirehub/EmployEase/job/impl/JobServiceImplementation.java
@@ -34,10 +34,12 @@ public class JobServiceImplementation implements JobService {
 
     @Override
     public boolean deleteJobById(Long id) {
-        try {
+        if(jobRepository.existsById(id))
+        {
             jobRepository.deleteById(id);
             return true;
-        } catch (Exception e) {
+        }
+        else{
             return false;
         }
     }
